[Fix] likeController 일부 수정

See merge request s03-webmobile1-sub3/s03p13a110!3

+++ backend/src/main/java/io/ssafy/p/i3a110/controller/CommentController.java
@@ -1,9 +1,11 @@
 package io.ssafy.p.i3a110.controller;
 
+import java.util.HashMap;
 import java.util.List;
 
 import javax.servlet.http.HttpSession;
 
+import org.apache.logging.log4j.message.Message;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -42,19 +44,24 @@ public class CommentController {
 			int uid = user.getId();
 			commentDto.setUid(uid);
 			commentService.addComment(commentDto);
-			return new ResponseEntity<>(null, HttpStatus.OK);
+			return new ResponseEntity<>(commentDto.getId(), HttpStatus.OK);
 		}else {
-			return new ResponseEntity<>(null, HttpStatus.BAD_REQUEST);
+			return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
 		}
 	}
 	
 	@DeleteMapping("/comments/{id}")
 	@ApiOperation(value = "댓글 삭제")
-	public void deleteComment(HttpSession session, @PathVariable String id) {
+	public Object deleteComment(HttpSession session, @PathVariable String id) {
 		String email = (String) session.getAttribute("email");
 		UserDto user = userService.findUserByEmail(email);
-		int uid = user.getId();
+		if(user!=null) {
+			int uid = user.getId();
+			commentService.deleteComment(id, uid);
+			return new ResponseEntity<>(HttpStatus.OK);
+		}else {
+			return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+		}
 		
-		commentService.deleteComment(id, uid);
 	}
 }
